requirements-dev.txt
-r requirements.txt

factory-boy==2.11.1
Faker==1.0.2
pytest==4.2.0
pytest-django==3.4.7
python-dateutil==2.8.0
text-unidecode==1.2
